requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
python-decouple==3.8
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
